requirements.txt
# Python 3.12.3 호환 버전
python-dotenv==1.0.1
openai==2.9.0
httpx==0.27.2
tiktoken==0.8.0
transformers==4.46.3
torch==2.5.1
numpy==1.26.4
trl==0.12.1
peft==0.13.2
wandb==0.18.7
accelerate==1.1.1
tqdm==4.67.1
pandas==2.2.3
scipy==1.14.1

# 선택적 패키지 (필요시 주석 해제)
# deeplake==3.9.28
# deepspeed==0.15.4
# bitsandbytes==0.45.0
# neural-compressor==3.1
# onnx==1.17.0